read xspress3-data
Created a new function read_nanomax_xspress3 from the read_nanomax_merlin function. Copied the openmultimerlin_roi function to the openmultixspress3_roi function and adjusted to the xspress3 detector.  Also, because we want to save two types of data, in the save_data function I added the detector name to the saved .mat filename