Added Pilatus opener, reading options, log
Added to Scan.m read_nanomax_pil100k that reads the data from a Pilatus detector. Also added the functionality of specifying if the scan is a stepscan (data_meta.scan.type = dmesh) or a flyscan (data_meta.scan.type = flyscan) in the meta data. It is required to specify scan.type for every reading. In the dmesh case, the dimensions of the scan (data_meta.scan.dim0, data_meta.scan.dim1) are also required. 

Added log functionality. Whenever a correction is made, it will be added to the property 'log' in Scan.